[cocos2d-objc] Fixed signature of SkeletonAnimation addAnimationForTrack, samples

+++ spine-cocos2d-objc/src/spine/SkeletonAnimation.m
@@ -179,7 +179,7 @@ static _TrackEntryListeners* getListeners (spTrackEntry* entry) {
 	return spAnimationState_setAnimation(_state, trackIndex, animation, loop);
 }
 
-- (spTrackEntry*) addAnimationForTrack:(int)trackIndex name:(NSString*)name loop:(bool)loop afterDelay:(int)delay {
+- (spTrackEntry*) addAnimationForTrack:(int)trackIndex name:(NSString*)name loop:(bool)loop afterDelay:(float)delay {
 	spAnimation* animation = spSkeletonData_findAnimation(_skeleton->data, [name UTF8String]);
 	if (!animation) {
 		CCLOG(@"Spine: Animation not found: %@", name);
